product featus added

+++ src/router/categoryRouter.js
@@ -2,6 +2,7 @@ import express from 'express'
 import slugify from 'slugify';
 import { updatCatValidation } from '../middlewares/joimiddleware.js';
 import { createNewCategory, deleteCat, readCategories, updateCategory } from '../model/category/CategoryModel.js';
+import { getSelectedProduct } from '../model/products/productsModel.js';
 
 const router = express.Router()
 
@@ -84,6 +85,17 @@ router.put("/", updatCatValidation, async(req, res, next )=>{
 //delete category
 router.delete("/:_id?", async (req, res, next )=>{
     const {_id} = req.params
+      //get all the product that as parentCat === _id
+  const prodList = await getSelectedProduct({parentCat: _id });
+  if (prodList.length) {
+    const names = prodList.map(({ name }) => name).toString();
+    return res.json({
+      status: "error",
+      message:
+        "Pelase re assign category for the following product before you can delete this category: " +
+        names,
+    });
+  }
     const result = await deleteCat(_id)
     
   if (result?._id) {
